hdis_queue_management/requirements.txt
requests
mysqlclient
Django==4.0.1
djangorestframework==3.13.1
django-oauth-toolkit==2.3.0
django-mysql==4.4.0
django-cors-headers==3.11.0
pika==1.2.0